Change encryption to Custom Casts
